safari-600.7-branch/LayoutTests/platform/mac/accessibility/aria-treegrid-expected.txt

Book Name / ISBN	Author	Price ($)
+ A Question of Love
978-3-453-40540-0	Nora Roberts	
9.99
This tests that the treegrid ARIA role does what it should (be exposed as a AXTable and support AXSelectedRows)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS treegrid.isAttributeSupported('AXSelectedRows') is true
PASS treegrid.selectedRowAtIndex(0).isEqual(selectedRow) is true
PASS selectedRow.isExpanded is false
PASS selectedRow.hierarchicalLevel is 0
PASS secondRow.isExpanded is true
PASS secondRow.hierarchicalLevel is 0
PASS thirdRow.isExpanded is true
PASS thirdRow.hierarchicalLevel is 1
PASS thirdRow.disclosedByRow().isEqual(secondRow) is true
PASS secondRow.disclosedRowAtIndex(0).isEqual(thirdRow) is true
PASS successfullyParsed is true

TEST COMPLETE
